+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `:menu` and `:menuitem` selector
 - Add `current:` filter for `:link` and `:link_or_button` selectors
 - Removed `focused:` in favour of the "native" capybara focus
+- Allow the `:modal` selector to select open `<dialog>` elements
 
 ## v0.8.2